feat: few updates based on fastapi documentation
