+++ apis/brand_api/crud.py
@@ -3,60 +3,50 @@ from uuid import UUID
 
 from sqlalchemy.orm import Session
 
-from .db.models import Brands, Categories, Users
-from .schemas import BrandsBase
+from . import schemas
+from .db.models import Brand, Category, User
 
 
-def create_brand(db: Session, brand: BrandsBase, user_id) -> Brands:
-    db_brand = Brands(
-        name=brand.name,
-        website=brand.website,
-        description=brand.description,
-        category_id=brand.category_id,
-        average_price=brand.average_price,
-        rating=brand.rating,
-        created_by=user_id,
-    )
+def create_brand(db: Session, brand: schemas.BrandsBase, user_id: UUID) -> Brand:
+    db_brand = Brand(**brand.dict(), created_by=user_id)
     db.add(db_brand)
     db.commit()
     db.refresh(db_brand)
     return db_brand
 
 
-def read_brand(db: Session, param: dict[str, str | UUID], show_deleted: bool = False) -> Brands:
+def read_brand(db: Session, param: dict[str, str | UUID], show_deleted: bool = False) -> Brand:
     filtering_param = list(param.keys())[0]
     return (
-        db.query(Brands)
+        db.query(Brand)
         .filter(
-            getattr(Brands, filtering_param, None) == param.get(filtering_param),
-            Brands.deleted_at == None if not show_deleted else Brands.deleted_at != None,
+            getattr(Brand, filtering_param, None) == param.get(filtering_param),
+            Brand.deleted_at == None if not show_deleted else Brand.deleted_at != None,
         )
         .first()
     )
 
 
-def read_all_brands(db: Session, skip: int = 0, limit: int = 100, show_deleted: bool = False) -> list[Brands]:
+def read_all_brands(db: Session, skip: int = 0, limit: int = 100, show_deleted: bool = False) -> list[Brand]:
     return (
-        db.query(Brands)
-        .filter(Brands.deleted_at == None if not show_deleted else Brands.deleted_at != None)
+        db.query(Brand)
+        .filter(Brand.deleted_at == None if not show_deleted else Brand.deleted_at != None)
         .offset(skip)
         .limit(limit)
         .all()
     )
 
 
-def update_brand(db: Session, brand) -> Brands:
+def update_brand(db: Session, brand) -> Brand:
     db.add(brand)
     db.commit()
     db.refresh(brand)
     return brand
 
 
-def create_category(db: Session, category, user_id) -> Categories:
-    db_category = Categories(
-        name=category.name,
-        description=category.description,
-        price_per_category=category.price_per_category,
+def create_category(db: Session, category: schemas.CategoriesBase, user_id: UUID) -> Category:
+    db_category = Category(
+        **category.dict(),
         created_by=user_id,
     )
     db.add(db_category)
@@ -65,29 +55,29 @@ def create_category(db: Session, category, user_id) -> Categories:
     return db_category
 
 
-def read_all_categories(db: Session, skip: int = 0, limit: int = 100, show_deleted: bool = False) -> list[Categories]:
+def read_all_categories(db: Session, skip: int = 0, limit: int = 100, show_deleted: bool = False) -> list[Category]:
     return (
-        db.query(Categories)
-        .filter(Categories.deleted_at == None if not show_deleted else Categories.deleted_at != None)
+        db.query(Category)
+        .filter(Category.deleted_at == None if not show_deleted else Category.deleted_at != None)
         .offset(skip)
         .limit(limit)
         .all()
     )
 
 
-def read_category(db: Session, param, show_deleted: bool = False) -> Categories:
+def read_category(db: Session, param, show_deleted: bool = False) -> Category:
     filtering_param = list(param.keys())[0]
     return (
-        db.query(Categories)
+        db.query(Category)
         .filter(
-            getattr(Categories, filtering_param, None) == param.get(filtering_param),
-            Categories.deleted_at == None if not show_deleted else Categories.deleted_at != None,
+            getattr(Category, filtering_param, None) == param.get(filtering_param),
+            Category.deleted_at == None if not show_deleted else Category.deleted_at != None,
         )
         .first()
     )
 
 
-def update_category(db: Session, category) -> Categories:
+def update_category(db: Session, category) -> Category:
     db.add(category)
     db.commit()
     db.refresh(category)
@@ -98,35 +88,35 @@ def update_user(db: Session, user) -> dict[str, bool]:
     db.add(user)
     db.commit()
     db.refresh(user)
-    return db.query(Users).filter(Users.id == user.id).first()
+    return db.query(User).filter(User.id == user.id).first()
 
 
-def read_user(db: Session, param, show_deleted: bool = False) -> Users:
+def read_user(db: Session, param: dict[str, str | UUID], show_deleted: bool = False) -> User:
     filtering_param = list(param.keys())[0]
     return (
-        db.query(Users)
+        db.query(User)
         .filter(
-            getattr(Users, filtering_param, None) == param.get(filtering_param),
-            Users.deleted_at == None if not show_deleted else Users.deleted_at != None,
+            getattr(User, filtering_param, None) == param.get(filtering_param),
+            User.deleted_at == None if not show_deleted else User.deleted_at != None,
         )
         .first()
     )
 
 
 def create_user(db: Session, user) -> dict[str, str]:
-    db_user = Users(email=user["email"], password=user["password"], created_at=datetime.now())
+    db_user = User(email=user["email"], password=user["password"], created_at=datetime.now())
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
     # TODO: This kind of return should probably be it's own CRUD function
-    test = db.query(Users).filter(Users.id == db_user.id).first()
+    test = db.query(User).filter(User.id == db_user.id).first()
     return test
 
 
 def read_all_users(db: Session, skip: int = 0, limit: int = 100, show_deleted: bool = False) -> list[dict[str, str]]:
     return (
-        db.query(Users)
-        .filter(Users.deleted_at == None if not show_deleted else Users.deleted_at != None)
+        db.query(User)
+        .filter(User.deleted_at == None if not show_deleted else User.deleted_at != None)
         .offset(skip)
         .limit(limit)
         .all()
@@ -135,7 +125,7 @@ def read_all_users(db: Session, skip: int = 0, limit: int = 100, show_deleted: b
 
 def read_one_user(db: Session, user_id, show_deleted: bool = False) -> dict[str, str]:
     return (
-        db.query(Users)
-        .filter(Users.id == user_id, Users.deleted_at == None if not show_deleted else Users.deleted_at != None)
+        db.query(User)
+        .filter(User.id == user_id, User.deleted_at == None if not show_deleted else User.deleted_at != None)
         .first()
     )
