+++ apis/brand_api/db/models.py
@@ -4,12 +4,12 @@ from datetime import datetime
 
 from sqlalchemy import Column, DateTime, Enum, ForeignKey, Integer, String
 from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import validates
+from sqlalchemy.orm import relationship
 
 from .database import Base
 
 
-class Brands(Base):
+class Brand(Base):
     __tablename__ = "brands"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
@@ -19,6 +19,8 @@ class Brands(Base):
     description = Column(String)
     average_price = Column(String)
     rating = Column(Integer)
+
+    category = relationship("Category", backref="brands")
 
     created_by = Column(UUID(as_uuid=True), ForeignKey("users.id", name="users_id_created_by"))
     updated_by = Column(UUID(as_uuid=True), ForeignKey("users.id", name="users_id_updated_by"), default=None)
@@ -37,7 +39,7 @@ class MyEnum(int, enum.Enum):
     five = 5
 
 
-class Categories(Base):
+class Category(Base):
     __tablename__ = "categories"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
@@ -54,7 +56,7 @@ class Categories(Base):
     deleted_at = Column(DateTime, default=None)
 
 
-class Users(Base):
+class User(Base):
     __tablename__ = "users"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
