+++ apis/brand_api/dependencies.py
@@ -8,7 +8,7 @@ from pydantic import ValidationError
 from sqlalchemy.orm import Session
 
 from .db.database import SessionLocal
-from .db.models import Users
+from .db.models import User
 from .schemas import SystemUser, TokenPayload
 
 reuseable_oauth = OAuth2PasswordBearer(tokenUrl="/login", scheme_name="JWT")
@@ -40,7 +40,7 @@ def get_current_user(db: Session = Depends(get_db), token: str = Depends(reuseab
             headers={"WWW-Authenticate": "Bearer"},
         )
 
-    user: Users = db.query(Users).filter(Users.email == token_data.sub).first()
+    user: User = db.query(User).filter(User.email == token_data.sub).first()
 
     if user is None:
         raise HTTPException(
